requirements.txt
vkbottle==4.4.1
openai-whisper
celery
redis
requests
python-dotenv
fastapi
uvicorn